Move promote and close channel actions to single menu
